pyproject.toml
[tool.poetry]
name = "temperatureagent"
version = "0.1.0"
description = "Temperature Alert Agent using fetchai's uagents library"
authors = ["Vaishnavi Anand <vaishnavi2010276@ssn.edu.in>"]
readme = "README.md"

[[tool.poetry.source]]
name="private"
url="https://github.com/Vaishnavi502/temperaturealertagent"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.13.0"
python-dotenv = "^1.0.0"
geocoder = "^1.38.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
